ai-engine/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
google-generativeai==0.7.2
pymongo==4.8.0
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.0
python-dotenv==1.0.1
